gke_autopilot/requirements.txt
tensorflow
streamlit
pillow
numpy
requests

#Run this file with - pip3 install -r requirements.txt